TabBarを追加しList表示とChart表示を切り替える処理を追加した
