+++ README.md
@@ -1,4 +1,4 @@
-# HealthKit の仕様調査のために作成した実験的プログラム
-## 歩数を表示する
-## 歩いた距離を表示する
-## 600m歩いた時の時間を表示する
+# HealthKitの仕様調査のための実験的プログラム
+## 1日の歩数を表示する
+## 1日で歩いた距離を表示する
+## 6分間で歩いた距離を表示する
